Pylint cleanup of posts_and_comments_bot.py Renamed Bow2.py to build_model.py
